error page and change id type

+++ src/main/java/com/ff/community/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable(name = "id") Integer id
+    public String publish(@PathVariable(name = "id") Long id
                           ,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -46,7 +46,7 @@ public class PublishController {
          @RequestParam("title") String title,
          @RequestParam("description") String description,
          @RequestParam("tag") String tag,
-        @RequestParam(name = "id",required = false) Integer id,
+        @RequestParam(name = "id",required = false) Long id,
         HttpServletRequest request,
         Model model) {
